feat: list pending and error states
