+++ src/containers/ListView/ListView.scss
@@ -1,0 +1,4 @@
+span.pgn__icon.breadcrumb-arrow {
+  width: 16px !important;
+  height: 16px !important;
+};
