+++ src/containers/ListView/index.jsx
@@ -3,22 +3,19 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 import {
-  DataTable,
-  TextFilter,
-  MultiSelectDropdownFilter,
   Container,
+  Spinner,
 } from '@edx/paragon';
-import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
-
-import { gradingStatuses } from 'data/services/lms/constants';
+import { FormattedMessage } from '@edx/frontend-platform/i18n';
 
 import { selectors, thunkActions } from 'data/redux';
-import lmsMessages from 'data/services/lms/messages';
+import { RequestKeys } from 'data/constants/requests';
 
-import StatusBadge from 'components/StatusBadge';
 import ReviewModal from 'containers/ReviewModal';
 
+import ListError from './ListError';
 import ListViewBreadcrumb from './ListViewBreadcrumb';
+import SubmissionsTable from './SubmissionsTable';
 import messages from './messages';
 import './ListView.scss';
 
@@ -29,142 +26,46 @@ export class ListView extends React.Component {
   constructor(props) {
     super(props);
     this.props.initializeApp();
-    this.handleViewAllResponsesClick = this.handleViewAllResponsesClick.bind(this);
-    this.selectedBulkAction = this.selectedBulkAction.bind(this);
-  }
-
-  get gradeStatusOptions() {
-    return Object.keys(gradingStatuses).map(statusKey => ({
-      name: this.translate(lmsMessages[gradingStatuses[statusKey]]),
-      value: gradingStatuses[statusKey],
-    }));
-  }
-
-  formatDate = ({ value }) => {
-    const date = new Date(value);
-    return date.toLocaleString();
-  }
-
-  formatGrade = ({ value: grade }) => (
-    grade === null ? '-' : `${grade.pointsEarned}/${grade.pointsPossible}`
-  );
-
-  formatStatus = ({ value }) => (<StatusBadge status={value} />);
-
-  translate = (...args) => this.props.intl.formatMessage(...args);
-
-  handleViewAllResponsesClick(data) {
-    const getsubmissionUUID = (row) => row.original.submissionUUID;
-    const rows = data.selectedRows.length ? data.selectedRows : data.tableInstance.rows;
-    this.props.loadSelectionForReview(rows.map(getsubmissionUUID));
-  }
-
-  selectedBulkAction(selectedFlatRows) {
-    return {
-      buttonText: this.translate(
-        messages.viewSelectedResponses,
-        { value: selectedFlatRows.length },
-      ),
-      className: 'view-selected-responses-btn',
-      handleClick: this.handleViewAllResponsesClick,
-      variant: 'primary',
-    };
   }
 
   render() {
-    // hide if submissions are not loaded.
-    if (this.props.listData.length === 0) {
-      return null;
-    }
-
+    const { isLoaded, hasError } = this.props;
     return (
       <Container className="py-4">
         <ListViewBreadcrumb />
-        <DataTable
-          isFilterable
-          numBreakoutFilters={2}
-          defaultColumnValues={{ Filter: TextFilter }}
-          isSelectable
-          isSortable
-          isPaginated
-          itemCount={this.props.listData.length}
-          initialState={{ pageSize: 10, pageIndex: 0 }}
-          data={this.props.listData}
-          tableActions={[
-            {
-              buttonText: this.translate(messages.viewAllResponses),
-              handleClick: this.handleViewAllResponsesClick,
-              className: 'view-all-responses-btn',
-              variant: 'primary',
-            },
-          ]}
-          bulkActions={[
-            this.selectedBulkAction,
-          ]}
-          columns={[
-            {
-              Header: this.translate(messages.username),
-              accessor: 'username',
-            },
-            {
-              Header: this.translate(messages.learnerSubmissionDate),
-              accessor: 'dateSubmitted',
-              Cell: this.formatDate,
-              disableFilters: true,
-            },
-            {
-              Header: this.translate(messages.grade),
-              accessor: 'points',
-              Cell: this.formatGrade,
-              disableFilters: true,
-            },
-            {
-              Header: this.translate(messages.gradingStatus),
-              accessor: 'gradingStatus',
-              Cell: this.formatStatus,
-              Filter: MultiSelectDropdownFilter,
-              filter: 'includesValue',
-              filterChoices: this.gradeStatusOptions,
-            },
-          ]}
-        >
-          <DataTable.TableControlBar />
-          <DataTable.Table />
-          <DataTable.EmptyTable content={this.translate(messages.noResultsFound)} />
-          <DataTable.TableFooter />
-        </DataTable>
+        { isLoaded && <SubmissionsTable /> }
+        { hasError && <ListError /> }
+        { (!isLoaded && !hasError) && (
+          <div className="w-100 h-100 text-center">
+            <Spinner animation="border" variant="primary" />
+            <h4><FormattedMessage {...messages.loadingResponses} /></h4>
+          </div>
+        )}
         <ReviewModal />
       </Container>
     );
   }
 }
 ListView.defaultProps = {
-  listData: [],
 };
 ListView.propTypes = {
-  // injected
-  intl: intlShape.isRequired,
   // redux
+  courseId: PropTypes.string.isRequired,
   initializeApp: PropTypes.func.isRequired,
-  listData: PropTypes.arrayOf(PropTypes.shape({
-    username: PropTypes.string,
-    dateSubmitted: PropTypes.number,
-    gradingStatus: PropTypes.string,
-    grade: PropTypes.shape({
-      pointsEarned: PropTypes.number,
-      pointsPossible: PropTypes.number,
-    }),
-  })),
-  loadSelectionForReview: PropTypes.func.isRequired,
+  isLoaded: PropTypes.bool.isRequired,
+  isPending: PropTypes.bool.isRequired,
+  hasError: PropTypes.bool.isRequired,
 };
 
 export const mapStateToProps = (state) => ({
-  listData: selectors.submissions.listData(state),
+  courseId: selectors.app.courseId(state),
+  isLoaded: selectors.requests.isCompleted(state, { requestKey: RequestKeys.initialize }),
+  isPending: selectors.requests.isPending(state, { requestKey: RequestKeys.initialize }),
+  hasError: selectors.requests.isFailed(state, { requestKey: RequestKeys.initialize }),
 });
 
 export const mapDispatchToProps = {
   initializeApp: thunkActions.app.initialize,
-  loadSelectionForReview: thunkActions.grading.loadSelectionForReview,
 };
 
-export default injectIntl(connect(mapStateToProps, mapDispatchToProps)(ListView));
+export default connect(mapStateToProps, mapDispatchToProps)(ListView);
