+++ src/data/redux/requests/index.js
@@ -1,1 +1,2 @@
 export { actions, reducer } from './reducer';
+export { default as selectors } from './selectors';
