+++ src/setupTest.js
@@ -21,6 +21,9 @@ jest.mock('@edx/frontend-platform/i18n', () => {
 });
 
 jest.mock('@edx/paragon', () => jest.requireActual('testUtils').mockNestedComponents({
+  Alert: {
+    Heading: 'Alert.Heading',
+  },
   AlertModal: 'AlertModal',
   ActionRow: 'ActionRow',
   Badge: 'Badge',
